refactor(org-server): use common-server error-handling
